+++ src/main/java/chess/view/InputView.java
@@ -1,8 +1,10 @@
 package chess.view;
 
+import chess.Command;
+import chess.domain.piece.PieceType;
+
 import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 import java.util.Scanner;
 
 public class InputView {
@@ -61,57 +63,34 @@ public class InputView {
         }
     }
 
-    public String requestPiece() {
+    public PieceType requestPiece() {
         System.out.println("Pawn이 Promotion 가능합니다.");
         System.out.println("[Quuen : q, Bishop : b, Knight : n, Rook : r]을 입력해주세요");
         String input = scanner.nextLine();
 
-        PromotionPiece.validate(input);
-
-        return input;
-    }
-
-    private enum Command {
-        START("start"),
-        END("end"),
-        MOVE("move"),
-        STATUS("status")
-        ;
-
-        private final String command;
-
-        Command(String command) {
-            this.command = command;
-        }
-
-        private static Command renderToCommand(String input) {
-            return Arrays.stream(values())
-                    .filter(value -> value.command.equals(input))
-                    .findAny()
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 명령어입니다."));
-        }
+        return PromotionPiece.renderToPieceType(input);
     }
 
     private enum PromotionPiece {
-        QUEEN("q"),
-        BISHOP("b"),
-        KNIGHT("n"),
-        ROOK("r");
+        QUEEN("q", PieceType.QUEEN),
+        BISHOP("b", PieceType.BISHOP),
+        KNIGHT("n", PieceType.KNIGHT),
+        ROOK("r", PieceType.ROOK);
 
         private final String command;
+        private final PieceType pieceType;
 
-        PromotionPiece(String command) {
+        PromotionPiece(String command, PieceType pieceType) {
             this.command = command;
+            this.pieceType = pieceType;
         }
 
-        private static void validate(String input) {
-            Optional<PromotionPiece> optionalPromotionPiece = Arrays.stream(values())
+        private static PieceType renderToPieceType(String input) {
+            return Arrays.stream(values())
                     .filter(value -> value.command.equals(input))
-                    .findAny();
-
-            if (optionalPromotionPiece.isEmpty()) {
-                throw new IllegalArgumentException("존재하지 않는 기물입니다.");
-            }
+                    .findFirst()
+                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 기물입니다."))
+                    .pieceType;
         }
     }
